requirements.txt
Django==2.0.6
django-enumchoicefield==1.0.0
django-use-email-as-username==1.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
PyJWT==1.6.4
pytz==2018.4
